feat: add function modules
